requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
docker==7.1.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
PyJWT==2.9.0
pywin32==308
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
waitress==3.0.2
